Spawn Tile Update
Added the SpawnTile. This Tile object uses the sprite Sprite.spawn. This
tile is the location at which the player is spawned. One of these tiles
is required to be in the map or an exception will be thrown. The Level
class reports to the Game class where the tile is located. The Game
class then created a Player object passing in the location of the
SpawnTile. The Player object spawns it self and then sets it location to
the location passed in. A multitude of changes had to occur in order to
make this happen. Most of these changes happened in the Level classes.
The large addition in the Level class is thanks to the findSpawn method.